+++ components/types.ts
@@ -9,8 +9,8 @@ export interface DeveloperProps {
   cohort: string;
   linkedin: string;
   password: string;
-  app: App; 
-  appId: number;
+  app?: App; 
+  appId?: number;
 }
 
 export interface DeveloperWithAppProps extends Developer {
@@ -31,8 +31,8 @@ export interface AppDataProps {
   appName: string,
   description: string,
   developers: Developer[],
-  appLink: string | null,
-  videoLink: string | null,
+  appLink?: string | null,
+  videoLink?: string | null,
   github: string,
   type: string,
   technologies: string[],
